Add: Checks for files and add warnings

+++ setup_environment.sh
@@ -15,21 +15,20 @@ if [ -f "$required_env_vars" ]; then
     echo "Getting required environment variables from $required_env_vars"
     echo "--------------------------------------------------------------"
     if [ ! -s "$required_env_vars" ]; then
-        echo "INFO: File $required_env_vars is empty continuing to process secrets"
+        echo "Warning: File $required_env_vars is empty continuing to process secrets"
     fi
     while IFS= read -r line || [[ -n "$line" ]]; do
         # Process each item in the list
-        echo "Fetching Variable: $line"
-        if [[ -v line ]]; then
-            : # Do nothing
+        echo "Fetching Local Enironmental Variable: $line"
+        if [[ -n ${!line} ]]; then
+        :
         else
-            echo "Environment variable $line does not exist or is not set. Set it with the command"
+            echo "ERROR: Environment variable $line does not exist or is not set. Set it with the command"
             echo "export $line=<$line>"
-            echo ""
             continue_flag=0
         fi
+        echo ""
     done < "$required_env_vars"
-    echo ""
 else
     continue_flag=0
     echo "File not found: $required_env_vars, Please create the file with the list of required environment variables."
@@ -43,10 +42,15 @@ if [[ continue_flag -eq 0 ]]; then
 else
     # Check if the file exists
     if [ -f "$secret_list" ]; then
-        rm $file_path
         # Read the file line by line and loop over the list
         echo "Getting secrets from $secret_list"
         echo "--------------------------------------------------------------"
+        if [ ! -s "$secret_list" ]; then
+            echo "Warning: File $secret_list has no secrets to fetch. if you want to fetch secrets, add them to the file and run the script again."
+            exit 1
+        fi
+        if [ -f "$file_path" ]; then rm $file_path; fi
+        
         while IFS= read -r line || [[ -n "$line" ]]; do
             # Process each item in the list
             echo "Fetching Secret: $line"
